+++ src/Node.js
@@ -293,10 +293,10 @@
                 maxY = Math.max(maxY, transformed.y);
             });
             return {
-                x: Math.round(minX),
-                y: Math.round(minY),
-                width: Math.round(maxX - minX),
-                height: Math.round(maxY - minY)
+                x: minX,
+                y: minY,
+                width: maxX - minX,
+                height: maxY - minY
             };
         },
         _drawCachedSceneCanvas: function(context) {