+++ test/unit/shapes/Ellipse-test.js
@@ -96,6 +96,6 @@ suite('Ellipse', function(){
         context.fill();
         context.lineWidth  = 8;
         context.stroke();
-        compareLayerAndCanvas(layer, canvas, 50);
+        compareLayerAndCanvas(layer, canvas, 80);
     });
 });