+++ src/plugins/Arrow.js
@@ -124,4 +124,3 @@
     Konva.Collection.mapMethods(Konva.Arrow);
 
 })();
-